+++ packages/templates/src/auth/templates/package-json.ts
@@ -5,20 +5,24 @@ export default {
   "version": "0.0.0",
   "private": true,
   "exports": {
-    ".": "./src/index.ts"
+    ".": {
+      "default": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
   },
   "scripts": {
+    "build": "tsc",
     "migrate": "dotenv -e .env -- auth migrate --config src/index.ts",
     "generate": "dotenv -e .env -- auth generate --config src/index.ts"
   },
   "dependencies": {
+    "auth": "^1.5.5",
     "better-auth": "^1.2.8",
     "pg": "^8.14.1"
   },
   "devDependencies": {
     "@<%= projectName %>/typescript-config": "workspace:*",
     "@types/pg": "^8.11.13",
-    "auth": "^1.5.5",
     "dotenv-cli": "^11.0.0",
     "typescript": "6.0.2"
   }