+++ packages/components/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @newt-app/components
 
+## 0.6.0
+
+### Patch Changes
+
+- @newt-app/ui@0.6.0
+
+## 0.5.0
+
+### Patch Changes
+
+- @newt-app/ui@0.5.0
+
+## 0.4.1
+
+### Patch Changes
+
+- @newt-app/ui@0.4.1
+
 ## 0.4.0
 
 ### Patch Changes